[NHUB-535] Convert settings module to async
