+++ newsroom/templates/settings.html
@@ -21,7 +21,7 @@
                         <a
                             data-test-id="settings-nav--{{ app._id }}"
                             class="side-navigation__btn {{ "active" if setting_type == app._id }}"
-                            href="{{ url_for('settings.app', app_id=app._id) }}"
+                            href="{{ url_for('settings.settings_app', app_id=app._id) }}"
                         >
                             {{ gettext(app.name) }}</a>
                     </li>
